refactor: move format_completion into completion.py
